blkparse: man: add absolute timestamp printing option
The functionality of printing out absolute timestamps has been
implemented in code but not documented in man pages.

When comparing the timings of related events with block I/O traces,
the absolute timestams play a key role.  I think that the
documentation of this might be beneficial to blktrace users.

The related commit was done in 2006 as follows,

> commit 7bd4fd0a4fca645bb50a641afac1e460a4e32dfd
> Author: Olaf Kirch <okir@lst.de>
> Date:   Fri Dec 1 10:34:11 2006 +0100
>
>     [PATCH] Add timestamp support
>
>     Signed-off-by: Jens Axboe <jens.axboe@oracle.com>
>

URL of the above patch,
https://git.kernel.org/pub/scm/linux/kernel/git/axboe/blktrace.git/commit/?id=7bd4fd0a4fca645bb50a641afac1e460a4e32dfd

Acked-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Hiroaki Mihara <hmihara@redhat.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>